Small change to fetchjson among other changes
Started writing the required file explaining some of the files purposes.

Utilizing nanostores to add a new comment to the comments list when posted without fetching the comments again.

if the success key is present in the returned json object, fetchJson now does not return null when the status is outside of 200